+++ home-helper-front/src/components/organisms/consume_point/ExchangePointContainer.tsx
