+++ home-helper-front/src/components/organisms/point_history/PointHistoryPresenter.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import { Tab } from "@components/atoms/Tab";
-import { ChargedPointHistoryContainer as ChargedPointHistory } from "@components/presenter/charged_point_history/ChargedPointHistoryContainer";
-import { UsedPointHistoryContainer as UsedPointHistory } from "@components/presenter/used_point_history/UsedPointHistoryContainer";
+import { ChargedPointHistoryContainer as ChargedPointHistory } from "@components/organisms/charged_point_history/ChargedPointHistoryContainer";
+import { UsedPointHistoryContainer as UsedPointHistory } from "@components/organisms/used_point_history/UsedPointHistoryContainer";
 
 type Props = {};
