+++ home-helper-front/src/components/organisms/possession_point/PossessionPointContainer.tsx
