+++ home-helper-front/src/components/organisms/possession_point/PossessionPointPresenter.tsx
