+++ home-helper-front/src/components/organisms/used_point_history/UsedPointHistoryContainer.tsx
