+++ home-helper-front/src/components/page/charge_point/ChargePointPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { ChargePointContainer } from "@components/presenter/charge_point/ChargePointContainer";
+import { ChargePointContainer } from "@components/organisms/charge_point/ChargePointContainer";
 import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
