+++ home-helper-front/src/components/page/home_helper/HomeHelperPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HomeHelperContainer } from "@components/presenter/home_helper/HomeHelperContainer";
+import { HomeHelperContainer } from "@components/organisms/home_helper/HomeHelperContainer";
 import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
