+++ home-helper-front/src/components/page/house_hold/HouseHoldPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HouseHoldContainer } from "@components/presenter/house_hold/HouseHoldContainer";
+import { HouseHoldContainer } from "@components/organisms/house_hold/HouseHoldContainer";
 import { HouseHoldLayout } from "@components/atoms/Layout/HouseHoldLayout";
 
 export const Page: FC = () => (
