+++ home-helper-front/src/components/page/prosession_point/PossessionPointPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { PossessionPointContainer } from "@components/presenter/possession_point/PossessionPointContainer";
+import { PossessionPointContainer } from "@components/organisms/possession_point/PossessionPointContainer";
 import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
