+++ home-helper-front/src/components/page/top/TopPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { TopContainer } from "@components/presenter/top/TopContainer";
+import { TopContainer } from "@components/organisms/top/TopContainer";
 import { TopLayout } from "@components/atoms/Layout/TopLayout";
 
 export const Page: FC = () => (
