+++ home-helper-front/src/components/page/use_point/ExchangePointPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { ExchangePointContainer } from "@components/presenter/consume_point/ExchangePointContainer";
+import { ExchangePointContainer } from "@components/organisms/consume_point/ExchangePointContainer";
 import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
